+++ src/main/java/lotto/model/InputValidator.java
@@ -5,14 +5,14 @@ import static lotto.common.AppConstant.LOTTO_NUMBER_COUNT;
 import static lotto.common.AppConstant.LOTTO_START_RANGE;
 import static lotto.common.AppConstant.LOTTO_UNIT_PRICE;
 import static lotto.common.AppConstant.SPLIT_DELIMITER;
-import static lotto.common.error.InputErrorType.BONUS_NUMBER_DUPLICATE_ERROR;
-import static lotto.common.error.InputErrorType.DIVIDED_BY_PRICE_ERROR;
-import static lotto.common.error.InputErrorType.LOTTO_PRICE_ERROR;
-import static lotto.common.error.InputErrorType.NEGATIVE_NUMBER_ERROR;
-import static lotto.common.error.InputErrorType.NUMBER_DUPLICATE_ERROR;
-import static lotto.common.error.InputErrorType.NUMBER_RANGE_ERROR;
-import static lotto.common.error.InputErrorType.PARSE_NUMBER_ERROR;
-import static lotto.common.error.InputErrorType.WINNING_NUMBER_SIZE_ERROR;
+import static lotto.common.AppErrorType.BONUS_NUMBER_DUPLICATE_ERROR;
+import static lotto.common.AppErrorType.DIVIDED_BY_PRICE_ERROR;
+import static lotto.common.AppErrorType.LOTTO_PRICE_ERROR;
+import static lotto.common.AppErrorType.NEGATIVE_NUMBER_ERROR;
+import static lotto.common.AppErrorType.NUMBER_DUPLICATE_ERROR;
+import static lotto.common.AppErrorType.NUMBER_RANGE_ERROR;
+import static lotto.common.AppErrorType.PARSE_NUMBER_ERROR;
+import static lotto.common.AppErrorType.WINNING_NUMBER_SIZE_ERROR;
 
 import java.util.Arrays;
 import java.util.HashSet;
